assignment-2-784915/code/mysimbdp-server/requirements.txt
flask 
flask-jsonpify 
flask-restful
cassandra-driver
numpy
pandas
argparse